+++ CHANGELOG.md
@@ -3,10 +3,16 @@
 ## Features
 
 - Camera touchscreen for focus
+- Camera focus and aperture are now animatable
+- Video output (Shot manager & Cameras)
+- Camera Snapshot
+- Collimator is now hidden when camera is not active
 
 ## Fixes
 
 - Lobby project list buttons states
+- Camera feedback is now always visible (even if no camera is active - only when it is activated)
+- Camera snapshot for cameras list when loading a scene
 
 # 0.1.0